chore: rename eslint.config.mjs to ____eslint.config.mjs
