+++ eslint.config.mjs
@@ -1,74 +1,65 @@
+import typescriptEslint from "@typescript-eslint/eslint-plugin";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import js from "@eslint/js";
 import { FlatCompat } from "@eslint/eslintrc";
-import eslint from "@eslint/js";
-import eslintConfigPrettier from "eslint-config-prettier";
-import eslintPluginPrettierRecommended from "eslint-plugin-prettier/recommended";
-import pluginSecurity from "eslint-plugin-security";
-import sonarjs from "eslint-plugin-sonarjs";
-import eslintPluginUnicorn from "eslint-plugin-unicorn";
-import globals from "globals";
-import * as tseslint from "typescript-eslint";
 
-const compat = new FlatCompat();
-
-const unicorn = {
-  languageOptions: {
-    ecmaVersion: 2024,
-    globals: {
-      ...globals.node
-    }
-  },
-  plugins: {
-    unicorn: eslintPluginUnicorn
-  },
-  rules: {
-    "unicorn/better-regex": "error",
-    "@typescript-eslint/consistent-type-imports": "error"
-  }
-};
-
-const storybookConfig = compat.config({
-  extends: [
-    "plugin:storybook/recommended",
-    "next",
-    "next/core-web-vitals",
-    "next/typescript"
-    // other extends
-    // eslintPluginStorybook.configs.recommended,
-  ],
-  rules: {
-    "storybook/no-uninstalled-addons": "off",
-    "storybook/hierarchy-separator": "off",
-    "storybook/prefer-pascal-case": "off",
-    "storybook/story-exports": "off",
-    "@next/next/no-duplicate-head": "off"
-  },
-  ignorePatterns: ["!.storybook", "storybook-static"]
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const compat = new FlatCompat({
+    baseDirectory: __dirname,
+    recommendedConfig: js.configs.recommended,
+    allConfig: js.configs.all
 });
-const eslintConfigs = [
-  eslint.configs.recommended,
-  ...tseslint.configs.recommended,
-  sonarjs.configs.recommended,
-  pluginSecurity.configs.recommended,
-  unicorn,
-  eslintPluginPrettierRecommended,
-  eslintConfigPrettier,
-  ...storybookConfig,
-  { files: ["*/**/main-stacks-section.tsx"], rules: { "@next/next/no-img-element": "off" } },
-  { files: ["tailwind.config.ts"], rules: { "@typescript-eslint/no-require-imports": "off" } },
 
-  {
-    ignores: [
-      ".config/*",
-      "build/",
-      "dist/",
-      ".xo-config.js",
-      ".*.js",
-      "node_modules",
-      "public",
-      ".next",
-      "CHANGELOG.md"
-    ]
-  }
+export default [
+    ...compat.extends("next/core-web-vitals", "next", "airbnb", "./.prettier.eslintrc.cjs"),
+    {
+        plugins: {
+            "@typescript-eslint": typescriptEslint,
+        },
+
+        rules: {
+            "@typescript-eslint/ban-ts-comment": "warn",
+            "react/react-in-jsx-scope": "off",
+            "react/jsx-filename-extension": "off",
+            "react/jsx-props-no-spreading": "off",
+            "import/extensions": "off",
+            "import/prefer-default-export": "off",
+            "no-shadow": "off",
+            "import/no-extraneous-dependencies": "off",
+            "import/no-amd":"off"
+        },
+    },
+    {
+        files: ["**/*.config.{ts,js}"],
+
+        rules: {
+            "global-require": "off",
+        },
+    },
+    {
+        files: ["**/*.test.{ts,tsx}"],
+
+        rules: {
+            "no-undef": "off",
+
+            "import/no-extraneous-dependencies": ["error", {
+                devDependencies: ["**/*.test.ts", "**/*.test.tsx"],
+            }],
+        },
+    },
+    {
+        ignores: [
+            ".config/*",
+            "build/",
+            "dist/",
+            ".xo-config.js",
+            ".*.js",
+            "node_modules",
+            "public",
+            ".next",
+            "CHANGELOG.md"
+        ]
+    }
 ];
-
-export default eslintConfigs;
